rfctr(html): move HtmlPartitionerOptions
Now that the broad refactoring is done and HTMLDocument is on its way
out, move `HtmlPartitionerOptions` to its natural home in
`unstructured.partition.html.partition` alongside `partition_html()`.
